Add files via upload
This is an HR Report created on real world data from a real company. 

First, I cleaned the data in excel. Then imported it in Power BI, and transformed it with DAX and Power Query. Put all the necessary formulas, and did all the required transformation to create this report.